@@ -3,106 +3,56 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@babel/core": "^7.1.2",
-    "autoprefixer": "^9.3.1",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.4",
-    "babel-plugin-add-react-displayname": "0.0.5",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-react-app": "^6.0.0",
-    "case-sensitive-paths-webpack-plugin": "1.1.4",
-    "chalk": "1.1.3",
-    "classnames": "^2.2.5",
-    "connect-history-api-fallback": "1.3.0",
-    "cross-spawn": "4.0.2",
-    "css-loader": "^1.0.1",
-    "detect-port": "1.0.1",
-    "dotenv": "2.0.0",
-    "elastic-apm-rum": "file:elastic-apm-rum",
-    "eslint": "^7.29.0",
-    "eslint-config-react-app": "^0.5.0",
-    "eslint-loader": "^4.0.2",
-    "eslint-plugin-flowtype": "5.8.2",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "2.2.3",
-    "eslint-plugin-react": "6.4.1",
-    "extract-text-webpack-plugin": "^2.1.2",
-    "file-loader": "0.9.0",
-    "filesize": "3.3.0",
-    "fs-extra": "0.30.0",
-    "gzip-size": "3.0.0",
-    "history": "^4.10.1",
-    "html-webpack-plugin": "2.24.0",
-    "http-proxy-middleware": "0.17.2",
-    "jest": "17.0.2",
-    "json-loader": "0.5.4",
-    "lerna": "^3.16.4",
-    "moment": "^2.17.1",
-    "object-assign": "4.1.0",
-    "path-exists": "2.1.0",
-    "postcss-loader": "^3.0.0",
-    "promise": "7.1.1",
-    "prop-types": "^15.7.2",
-    "react": "^16.2.0",
-    "react-dev-utils": "^0.4.2",
-    "react-dom": "^16.2.0",
-    "react-redux": "^5.0.1",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "recursive-readdir": "2.1.0",
-    "redux": "^3.6.0",
-    "redux-thunk": "^2.1.0",
-    "strip-ansi": "3.0.1",
-    "style-loader": "^0.13.1",
-    "uglifyjs-webpack-plugin": "^1.3.0",
-    "url-loader": "0.5.7",
-    "webpack": "^3.12.0",
-    "webpack-dev-server": "1.16.2",
-    "webpack-manifest-plugin": "1.1.0",
-    "whatwg-fetch": "1.0.0"
+    "@elastic/apm-rum": "^5.12.0",
+    "@elastic/apm-rum-react": "^1.4.2",
+    "@reduxjs/toolkit": "^1.9.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "classnames": "^2.3.2",
+    "http-proxy-middleware": "^2.0.6",
+    "jest": "29.3.1",
+    "moment": "^2.29.4",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.7.0",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
+    "redux-thunk": "^2.4.2",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "preinstall": "if ! [ -d elastic-apm-rum ]; then (git clone https://github.com/elastic/apm-agent-rum-js elastic-apm-rum --depth 1; cd elastic-apm-rum; git fetch origin '@elastic/apm-rum@5.11.0'; git reset --hard FETCH_HEAD) fi",
-    "postinstall": "npm install --prefix ./elastic-apm-rum --dev --prod && npx lerna run build --stream && rm -rf ./elastic-apm-rum/node_modules",
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "test": "node scripts/test.js --env=jsdom"
+    "test": "node scripts/test.js --env=jsdom",
+    "mock-api-server": "node scripts/mock-server.js",
+    "install-dev": "npm install --also=dev",
+    "build": "react-scripts build",
+    "start": "react-scripts start",
+    "test-react": "react-scripts test",
+    "build-analyze": "npm run build && npm run analyze"
   },
   "proxy": "http://localhost:3001/",
-  "jest": {
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx}"
-    ],
-    "setupFiles": [
-      "<rootDir>/config/polyfills.js"
-    ],
-    "testPathIgnorePatterns": [
-      "<rootDir>[/\\\\](build|docs|node_modules)[/\\\\]"
-    ],
-    "testEnvironment": "node",
-    "testURL": "http://localhost",
-    "transform": {
-      "^.+\\.(js|jsx)$": "<rootDir>/node_modules/babel-jest",
-      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-      "^(?!.*\\.(js|jsx|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
-    },
-    "transformIgnorePatterns": [
-      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx)$"
-    ],
-    "moduleNameMapper": {
-      "^react-native$": "react-native-web"
-    }
-  },
-  "babel": {
-    "presets": [
-      "react-app"
-    ],
-    "plugins": [
-      "add-react-displayname"
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
     ]
   },
-  "eslintConfig": {
-    "extends": "react-app"
+  "devDependencies": {
+    "npm-check-updates": "^16.6.2",
+    "react-dev-utils": "^12.0.1"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
