@@ -1,13 +1,10 @@
 import React, { Component } from 'react';
 import classnames from 'classnames';
-import { Route, Redirect, Switch } from 'react-router-dom'
-import { ApmRoute } from '@elastic/apm-rum-react'
-
-import './style.css';
+import { useParams,Route, Navigate, Routes } from 'react-router-dom'
+import { withTransaction } from '@elastic/apm-rum-react'
 
 import AppHeader from '../AppHeader';
 import AppFooter from '../AppFooter';
-
 import Dashboard from '../Dashboard';
 import Product from '../Product';
 import Products from '../Products';
@@ -17,6 +14,50 @@ import Customers from '../Customers';
 import Customer from '../Customer';
 import NotFound from '../NotFound';
 
+import './style.css';
+
+// Hooks can only be called inside the body of a function component.
+function CustomerId() {
+  let { id } = useParams()
+  return <Customer customerId={id} />
+}
+
+function OrderId() {
+  let { id } = useParams()
+  return <Order orderId={id} />
+}
+
+function ProductId() {
+  let { id } = useParams()
+  return <Product productId={id} />
+}
+
+/*
+react-router-dom v6 is not supported https://github.com/elastic/apm-agent-rum-js/issues/1176
+<ApmRoute path="/products" component={Products} />
+so we use withTransaction instead
+*/
+
+function OrdersWithTransaction(props) {
+  return withTransaction('Orders', 'Component')(Orders)
+}
+
+function CustomersWithTransaction(props) {
+  return withTransaction('Customers', 'Component')(Customers)
+}
+
+function CustomerWithTransaction(props) {
+  return withTransaction('Customer', 'Component')(Customer)
+}
+
+function OrderWithTransaction(props) {
+  return withTransaction('Order', 'Component')(Order)
+}
+
+function ProductWithTransaction(props) {
+  return withTransaction('Product', 'Component')(Product)
+}
+
 class App extends Component {
 
   render() {
@@ -24,29 +65,22 @@ class App extends Component {
     return (
       <div className={classnames('App', className)}>
         <AppHeader />
-
         <div className="AppMain">
-          <Switch>
-            <ApmRoute
-              exact
-              path="/"
-              component={() => (
-                <Redirect
-                  to={{
-                    pathname: '/dashboard'
-                  }}
-                />
-              )}
-            />
-            <ApmRoute path="/dashboard" component={Dashboard} />
-            <ApmRoute path="/products" component={Products} />
-            <ApmRoute path="/products/:id" component={Product} />
-            <ApmRoute path="/orders" component={Orders} />
-            <ApmRoute path="/orders/:id" component={Order} />
-            <ApmRoute path="/customers" component={Customers} />
-            <ApmRoute path="/customers/:id" component={Customer} />
-            <Route component={NotFound} />
-          </Switch>
+        <Routes>
+          <Route
+                exact
+                path="/"
+                element={<Navigate to='/dashboard' />}
+              />
+          <Route path="/dashboard" element={<Dashboard/>} />
+          <Route path="/products" element={<Products />} />
+          <Route path="/products/:id" element={<ProductId />} />
+          <Route path="/orders" element={<Orders />} />
+          <Route path="/orders/:id" element={<OrderId />} /> 
+          <Route path="/customers" element={ <Customers />} />
+          <Route path="/customers/:id" element={<CustomerId />} />
+          <Route element={NotFound} />
+        </Routes>
         </div>
 
         <AppFooter />
