@@ -1,4 +1,4 @@
-import {createStore, applyMiddleware, compose} from 'redux';
+import {createStore, applyMiddleware, compose} from '@reduxjs/toolkit';
 import rootReducer from '../reducers/rootReducer';
 import thunk from 'redux-thunk';
 
